requirements.txt
colorama
google-cloud-translate
